Cleanup: use integer argument where useful and promote to double internally.
git-svn-id: svn://svn.code.sf.net/p/bogofilter/code/trunk/bogofilter@5001 046ba1d7-8659-4f4f-8137-08d775d845e8

+++ src/prob.c
@@ -15,7 +15,7 @@ AUTHORS:
 #include "prob.h"
 
 double calc_prob_pure(uint good, uint bad,
-	double goodmsgs, double badmsgs,
+	uint goodmsgs, uint badmsgs,
 	double s, double x)
 {
     int n = good + bad;
@@ -36,7 +36,8 @@ double calc_prob_pure(uint good, uint bad,
 	*/
 
 	/* This modified version, with 1 division, is considerably% faster. */
-	pw = bad * goodmsgs / (bad * goodmsgs + good * badmsgs);
+	pw =   bad * (double)goodmsgs
+	    / (bad * (double)goodmsgs + good * (double)badmsgs);
 
 	fw = (s * x + n * pw) / (s + n);
     }
